LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/browsers/the-window-object/noopener-noreferrer-sizing.window-expected.txt


FAIL window.open() with noopener should have equal viewport width and height assert_equals: expected 980 but got 800
FAIL window.open() with noreferrer should have equal viewport width and height assert_equals: expected 980 but got 800
